moved faq.module.scss to /pages
