+++ src/pages/faq.module.scss
